### 2024-11-07 #7439 @warrickball
Hi @evavagiakis, thanks for your submission! Given the small size of this code base, I'm going to submit your article for editorial review on whether it meets JOSS's requirements for ["substantial scholarly effort"](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). This usually takes a week or two and I'll report back here as soon as I hear back from the rest of the editorial board.

Thanks for your patience!
