### 2024-11-08 #7439 @warrickball
@evavagiakis — Would you mind fixing Samantha Usman's ORCID, which is currently causing the compilation failure for the paper? Once you've done so, you can comment with `@editorialbot generate pdf` to try again. It'll be easier for editors to assess the scope of this submission if we can compile the paper.
