### 2024-11-16 #7439 @warrickball
Hi @evavagiakis ­— The editorial team has agreed that this is suitable for JOSS review, so we'll proceed as usual. For now that means waiting for an editor to become available, so I'll place this on our waiting list. Thanks for your patience!
