### 2025-01-10 #7439 @evavagiakis
Hi @JBorrow, thank you! Some recommendations from our team are Kimmy Wu (wlwu at slac.stanford.edu), Sebastian Bocquet (sebastian.bocquet at physik.lmu.de), and Zack Li (zackli at berkeley.edu)
