### 2025-01-29 #7439 @JBorrow
Hi @weiguangcui, I think you would be an excellent reviewer for this paper and package for JOSS. We are a Diamond Open Access journal focused on research software, and our reviewers guide is here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. Would you be available as a reviewer?
