### 2025-02-11 #7439 @JBorrow
@avapolzin great, thank you for disclosing the CoI too. Can you take a look at our policy (https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) and let us know, e.g. if "you and the submitter are both employed by the same very large organization but in different units without any knowledge of each other."
