### 2025-02-11 #7439 @avapolzin
Hi @JBorrow, we're nominally in the same department, but have no research collaborations/overlap, so I can make an impartial assessment of the submitted work. (Samantha Usman and I have worked together at the University's Center for Teaching and Learning.)
