### 2025-02-12 #7439 @JBorrow
Fantastic, thank you. I have spoken with other editors and we are happy to waive this potential CoI.
