### 2025-02-12 #7439 @JBorrow
We now have two reviewers! The review will actually take place in another thread, created shortly. Please let me know if you have any questions, and don't forget to check out our reviewer's guide here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html.
