### 2025-02-12 #7774 @editorialbot
[REVIEW]: DeepSZSim: A Python code for fast, tunable Sunyaev–Zeldovich effect submap simulations
**Submitting author:** @evavagiakis (<a href="http://orcid.org/0000-0002-2105-7589">Eve Vavagiakis</a>)
**Repository:** https://github.com/deepskies/DeepSZSim
**Branch with paper.md** (empty if default branch): main
**Version:** v0.1.1
