### 2025-03-17 deepskies/DeepSZSim#241 @avapolzin
[JOSS Review] Installation, documentation, issues/contributing, manuscript content
Hi, everyone,

Here are some initial comments on the installation instructions, documentation, issues/contributing, and manuscript content as part of the JOSS review.

**Installation:**
- Is there a reason that the pip installation from PyPI is not listed in the README? As it exists and may be preferred by some users, it should be included in the installation instructions. If the concern is with environment creation or direct dependencies, this can be stated in the documentation, but right now there’s an orphaned PyPI entry.
- Within the pyproject.toml file, it looks like an old email is used for the first author. Assuming that email is still accessible/checked, this is not a problem, but it should be updated if it won’t be monitored.

**Documentation:**
- The README is very sparse, and it will need to be fleshed out, as the Read the Docs page points to it. At the same time, the Read the Docs should be standalone and need to have sufficiently detailed information to serve as documentation (otherwise users could just look at your docstrings). Generating a README-like .rst file for sphinx is relatively trivial, and it would also be good to include usage examples or similar in the official documentation.
- At no point is the Read the Docs page listed in the available README — much like the absent PyPI reference despite an existing record, it will be a confusing omission for users. Please link directly to the Read the Docs page from your README (the Read the Docs badges are very nice for this and offer a check that the documentation compiled properly).
- Within deepszsim.readthedocs.io, some of the docstrings were not read in properly by sphinx. Please check the formatting to ensure that it is uniform and readable. Similarly, some of the functions are missing descriptions. Please add those to each function regardless of whether they are user-facing or not per PEP 8.

**Issues/Contributions:**:
- As there are quite a few authors, it would be appropriate to be clear about who exactly should be contacted about contributions. If it is, in fact, all of the authors, it should be made explicit. If there are authors who are specifically charged with maintaining the software, please list explicitly who users should contact as this will facilitate engagement. Currently, only some authors even have websites linked, so please also include emails (it can be in name [at] institution [dot] edu format) for the relevant contacts.
- There should be a formalized procedure for bug reports, too. Many people will default to opening an issue, but please also be clear about whether you want direct contact (i.e.,  emails — and where they should go) or if issues are sufficient.
- There are many open issues in the repository currently — a quick glance through suggests that many of these should be closed as completed. Maintainers should go through and close completed issues and address any open ones that are pressing and critical to the project. This can exclude wishlist items, but with a substantial backlog of open issues, it will appear like the project is not being actively maintained. (If you would like a forum to publicly engage with authors about improvements to the code, consider using the discussions feature on GitHub to avoid muddying the purpose of issues by leaving so many open.)

**Paper/Paper Draft:**
- I’m sure this will be addressed by the editors, but some of the references in the paper are not processed properly in the resultant PDF (from “DeepCMBSim Module” on) because you have enclosed the references in backticks. This will be trivial to address.
- On occasion, in-text references are used when the citations should be parenthetical and vice versa. The paper should be proofread to fix these instances.
- Line 82, “Figure Figure 2”.
- It does not appear that `DeepCMBSim` has been referenced (whether formally or via footnote linking to the parent repo). As this is the basis of your code, it is critical that this properly credited.
- CAMB should be cited on first reference.
- Line 89, M_{200}.
- Line 101, "Figure Figure 3".

I will follow up soon with comments about the code and functionality.
