### 2025-03-17 deepskies/DeepSZSim#242 @avapolzin
[JOSS Review] Code and functionality
Hi, everyone,

Here are my comments on the code and its functionality.

**Tests:**
- Tests are not automated and, though they exist, there is not any instruction for how users should run them. Since the code is explicitly intended to be easy to use for beginners and students, this should be listed — even a simple `python -m pytest tests/*` is sufficient. I did run the tests myself, and they all pass.

**Examples:**
- Examples are in Jupyter notebook/lab. Neither is installed by default, though they should be so that users can actually run the examples. Similarly, there are not instructions for running the notebooks with a kernel generated for the environment. This should be included explicitly with instructions to make sure users have how to access the examples.
```bash
pip install notebook
conda activate szsims
python -m ipykernel install --user --name szsims --display-name "deepszsim"
cd notebooks
Jupyter notebook
```
- I ran the example notebooks straight through and got several errors and had output plots that looked nothing like the original example. I believe the code structure/functions have changed since the notebooks were created. Given that these are the examples, it should be possible to run them straight through without errors. This **must** be addressed. Especially since the notebooks currently serve as documentation, this is absolutely critical.
- The example in the README proceeds as expected.

**General Repository:**
- Please remove the extraneous .egg_info, .DS_Store, etc. files. These should be added to the .gitignore with wildcards specifying the path.
