### 2025-03-17 #7774 @avapolzin
Hi, everyone, I have provided comments in two issues in the `DeepSZSim` repository: [#241](https://github.com/deepskies/DeepSZSim/issues/241) and [#242](https://github.com/deepskies/DeepSZSim/issues/242). The first addresses the content of the installation instructions, documentation, manuscript, and issues/contributing guidelines, and the second addresses the code and package functionality. My primary concern is that the examples in DeepSZSim/notebooks do not work on a straight run through due to errors that indicate that the code itself was changed. Since there is very minimal documentation that points to these notebooks as the primary means of users understanding the code, this absolutely critical to address.
