### 2025-03-17 deepskies/DeepSZSim#242 @avapolzin
PDFs of the example notebooks showing outputs if run straight through.
[demo_full_pipeline.pdf](https://github.com/user-attachments/files/19298753/demo_full_pipeline.pdf)
[demo_simulation.pdf](https://github.com/user-attachments/files/19298754/demo_simulation.pdf)
