### 2025-04-06 fca5aee
updated README for comments; updated environment.yml for numpy version control; updated notebook for compatibility
