+++ README.md
@@ -7,15 +7,31 @@ Code for producing fast simulations of the SZ effect for galaxy halos of varying
 The code is structured as depicted here: ![DeepSZSim workflow](paper/figures/DeepSZSim_Workflow.png)
 The CMB simulations are handled by [DeepCMBSim](https://www.github.com/deepskies/deepcmbsim), based on [CAMB](https://camb.info), and further by [pixell](https://github.com/simonsobs/pixell). The SZ cluster simluations are done in `make_sz_cluster.py` and instrumental effects are added in `filters.py` and `noise.py`.  
 
+Documentation is available via [readthedocs](https://deepszsim.readthedocs.io/en/latest/).
+
 ## Quickstart
 
 ### Installation 
 
-We provide an environment specification file for `conda` or `mamba` users at `environment.yml`, which will produce a new virtual environment called `szsims` with appropriate versions of major python packages. With `conda`, this environment is created by `conda env create -f environment.yml`. With `micromamba` the `env` is omitted and a new environment is instead created with `micromamba create -f environment.yml`.
+#### From `pypi`
+
+The simplest way to install `deepszsim` is via PyPI, using
+```commandline
+pip install deepszsim
+```
+The project history and source files are available on PyPI [here](https://pypi.org/project/deepszsim/).
+
+#### From Source
+
+We provide an environment specification file for `conda` or `mamba` users at `environment.yml`, which will produce a new virtual environment called `szsims` with appropriate versions of major python packages, after which you can install with `pip`. With `conda`, the workflow to create the environment, activate it, and install the package is 
+```
+conda env create -f environment.yml
+conda activate szsims
+pip install .
+```
+(With `micromamba` the `env` is omitted and a new environment is instead created with `micromamba create -f environment.yml`)
 
 The simulated CMB signal relies on `camb` and `pixell`, cosmology relies on `colossus`, and utilities for saving rely on `h5py`. These are specified in the `pyproject.toml` file.
-
-From the top-level directory, you can do `pip install .` to install the package.
 
 ### Usage
 
@@ -24,6 +40,10 @@ The usage of this code is documented in `notebooks/demo_simulation.ipynb`. A det
 A full list of potential inputs is documented in `settings/config.yaml` and you can edit `settings/inputdata.yaml` to reflect your desired simulation settings.  
 
 `dm_halo_dist.py` generates an array of mass and redshift. The functions in `make_sz_cluster.py` create pressure profiles, Compton-y, and SZ signal maps for a halo of a given mass and redshift, and produces the final simulated submaps. These submaps contain simulated CMB and simple instrument beam convolution from `simtools.py` and white noise from `noise.py`. Plotting tools are provided in `visualization.py`. Simulations of a large suite of clusters can be achieved easily with `simclusters.py`.
+
+### Tests
+
+Tests are provided in the `tests` directory. They are automatically run via CircleCI on pushes to the repository. Users can verify tests with `python -m pytest tests/*`.
 
 ### Example
 
@@ -47,7 +67,7 @@ If you use this code in your research, please cite this GitHub repo and our JOSS
 
 ## Contributing
 
-If you would like to contribute, please open a new [issue](https://github.com/deepskies/deepszsim/issues), and/or be in touch with the [authors](#contact)
+If you would like to contribute, you find a bug, or you have a feature request, please open a new [issue](https://github.com/deepskies/deepszsim/issues), and/or be in touch with the [authors](#contact).
 
 ## Contact
 
