### 2025-04-28 be7b5f5
updated README for comments; pyproject.toml for `notebook`; .gitignore for a hidden files
