+++ pyproject.toml
@@ -17,6 +17,7 @@ dependencies = [
   "colossus >= 1.3.0",
   "tqdm >= 4.66.0",
   "pixell == 0.17.3",
+  "notebook>=7.2.2"
 ]
 
 [tool.dev-dependencies]
