### 2025-05-06 deepskies/DeepSZSim#241 @samueldmcdermott
we are working on this now @avapolzin! One quick question -- you say

> Within deepszsim.readthedocs.io, some of the docstrings were not read in properly by sphinx. Please check the formatting to ensure that it is uniform and readable. Similarly, some of the functions are missing descriptions. Please add those to each function regardless of whether they are user-facing or not per PEP 8.

but things look ok to me. Can you provide a link to a particular example and I can go from there? thank you!
