### 2025-05-06 deepskies/DeepSZSim#241 @avapolzin
Hi, @bnord -- I think it will be a "cleaner" look to move wish list items to Discussions. Whether you want to consolidate them or not may rest on the timeline you foresee/whether some of them are related to each other; my feel is that should be determined by whatever works best for the collaboration/organization.
