### 2025-05-06 25df4e6
updated notebooks and comments in visualization.py
