### 2025-05-06 6e0112a
setting seed
