### 2025-05-14 6e4cd9b
Adding contact authors to paper.md
