### 2025-05-14 70e756b
Update io_params.py with filler docstring
