### 2025-05-25 c916dba
docstring formatting
