### 2025-05-25 1da358a
add top-level descriptor to docstrings
