### 2025-05-25 32e45e0
make quote types uniform
