### 2025-05-25 2133fec
add top-level descriptors to docstrings
