### 2025-05-25 0fd9ed6
consistent docstring quote type
