### 2025-05-25 b20fa15
consistent docstring quote types
