### 2025-05-25 797c4a6
docstring cleanup text
