+++ deepszsim/filters.py
@@ -11,20 +11,21 @@ def get_tSZ_signal_aperture_photometry(dT_map, radmax_arcmin, pixel_scale,
     
     Parameters:
     ----------
-    dT_map: array to represent the map in uK
+    dT_map: array(float)
+        map in uK
     radmax_arcmin: float
-        the radius of the inner aperture, in arcmin
+        radius of the inner aperture in arcmin
     pixel_scale: float
-        How many arcmin per pixel for the current settings
+        arcmin per pixel for the current settings
     fmax: float
-        fmax * radmax_arcmin is the radius of the outer aperture, in arcmin
+        fmax * radmax_arcmin is the radius of the outer aperture in arcmin
 
     Returns:
     -------
     disk_mean: float
-        The average value within an annulus of inner radius R
+        average value within an annulus of inner radius R
     ring_mean: float
-        The average value within an annulus of outer radius sqrt(2)*R
+        average value within an annulus of outer radius sqrt(2)*R
     tSZ signal: float
         thermal SZ effect signal
     """
