### 2025-05-25 eb1bc82
clean up docstrings
