+++ deepszsim/io_params.py
@@ -15,12 +15,12 @@ class config_obj:
     Attributes
     ----------
     CAMBparams : CAMBparams instance
-        this is necessary for CAMB to return results
+        CAMB object; returns results from CAMB
     UserParams : dict
-        a dictionary of values that the user has specified (smaller than the corresponding
+        dictionary of values that the user has specified (smaller than the corresponding
         dictionary that would be necessary to fully specify a CAMBparams instance)
     dict_iterables : dict
-        a dictionary of all of the iterables that the user has specified, which will be made
+        dictionary of all of the iterables that the user has specified, which will be made
         available to loop over in camb_power_spectrum.CAMBPowerSpectrum
     """
     def __init__(
@@ -65,7 +65,7 @@ class config_obj:
 
     def update_val(self, attr, new_val):
         """
-        updates values in the config_obj
+        Updates values in the config_obj
         
         Parameters
         ----------
@@ -73,6 +73,10 @@ class config_obj:
             a key of the UserParams dictionary
         new_val : float
             new value that you wish attr to take on
+
+        Returns
+        -------
+            none
         """
         attr_split = re.split("\.", attr)
         if attr in self._all_params_dict['USERPARAMS']:
@@ -94,6 +98,10 @@ class config_obj:
             a key of the UserParams dictionary
         new_val : float
             new value that you wish attr to take on
+
+        Returns
+        -------
+            none
         """
         with open(os.path.join(savedir, f"{run_id}_params.yaml"), permission) as yaml_file:
             dump = pyyaml.dump(self.dict, default_flow_style = False, allow_unicode = True, encoding = None)
@@ -102,13 +110,17 @@ class config_obj:
             
     def _generate_run_id(random_digits=6):
         """
-        <>
+        Generates run id; typically used in filenames
+        
         Parameters
         ----------
-        attr : str
-            a key of the UserParams dictionary
-        new_val : float
-            new value that you wish attr to take on
+        random_digits: int
+            exponent for the number of random digits in the seed
+
+        Returns
+        -------
+        runid: int
+            run identification number
         """
 
         _rint = np.random.randint(10**random_digits)
