### 2025-05-25 28830b0
clarify text in docstring
