+++ deepszsim/make_sz_cluster.py
@@ -10,11 +10,15 @@ from colossus.halo import mass_adv
 from astropy import constants as c
 from astropy import units as u
 
+# constants
 keVcm3_to_Jm3 = ((1 * u.keV / (u.cm**3.)).to(u.J / (u.m**3.))).value
-thermal_to_electron_pressure = 1 / 1.932  # from Battaglia 2012, assumes
 Mpc_to_m = (1 * u.Mpc).to(u.m).value
 Thomson_scale = (c.sigma_T/(c.m_e * c.c**2)).value
-# fully ionized medium
+
+# conversion factor from Battaglia 2012
+thermal_to_electron_pressure = 1 / 1.932  
+
+
 
 def _param_Battaglia2012(A0, alpha_m, alpha_z, M200_SM, redshift_z):
     """
@@ -30,15 +34,15 @@ def _param_Battaglia2012(A0, alpha_m, alpha_z, M200_SM, redshift_z):
     alpha_z: float
         power law index for the redshift dependent part
     M200_SM: float
-        the mass of the cluster at 200 times the critical density of the 
+        mass of the cluster at 200 times the critical density of the 
         universe in units of solar masses
     redshift_z: float
-        the redshift of the cluster
+        redshift of the cluster
 
     Returns:
     --------
     A: float
-        the parameter A given the formula from Battaglia 2012, Eq 11
+        formula from Battaglia 2012, Eq 11
     """
     
     A = A0 * (M200_SM / 1e14)**alpha_m * (1. + redshift_z)**alpha_z
@@ -54,15 +58,15 @@ def _P0_Battaglia2012(M200_SM, redshift_z):
     Parameters:
     -----------
     M200_SM: float
-        the mass of the cluster at 200 times the critical density of the 
+        mass of the cluster at 200 times the critical density of the 
         universe, in units of solar masses
     redshift_z: float
-        the redshift of the cluster (unitless)
+        redshift of the cluster (unitless)
 
     Returns:
     -------
     P0: float
-        the normalization factor for the Battaglia 2012 profile
+        normalization factor for the Battaglia 2012 profile
 
     """
     return _param_Battaglia2012(18.1, 0.154, -0.758, M200_SM, redshift_z)
@@ -76,15 +80,15 @@ def _xc_Battaglia2012(M200_SM, redshift_z):
     Parameters:
     -----------
     M200_SM: float
-        the mass of the cluster at 200 times the critical density of the 
+        mass of the cluster at 200 times the critical density of the 
         universe, in units of solar masses
     redshift_z: float
-        the redshift of the cluster (unitless)
+        redshift of the cluster (unitless)
 
     Returns:
     --------
     xc: float
-        the core-scale factor for the Battaglia 2012 profile
+        core-scale factor for the Battaglia 2012 profile
 
     """
     return _param_Battaglia2012(0.497, -0.00865, 0.731, M200_SM, redshift_z)
@@ -92,21 +96,20 @@ def _xc_Battaglia2012(M200_SM, redshift_z):
 
 def _beta_Battaglia2012(M200_SM, redshift_z):
     """
-    Calculates beta, the power law index, 
-    from Battaglia 2012, using the values from Table 1,
+    Calculates beta, the power law index, from Battaglia 2012 from Table 1
 
     Parameters:
     ----------
     M200_SM: float
-        the mass of the cluster at 200 times the critical density of the 
+        mass of the cluster at 200 times the critical density of the 
         universe, in units of solar masses
     redshift_z: float
-        the redshift of the cluster (unitless)
+        redshift of the cluster (unitless)
 
     Returns:
     -------
     beta: float
-        the power law index for the Battaglia 2012 profile
[…]
