+++ deepszsim/noise.py
@@ -8,7 +8,8 @@ def generate_noise_map(image_size, noise_level, pix_size):
     """
     Generates a white noise map based on the noise level and beam size.
 
-    Args:
+    Parameters:
+    -----------
         image_size: int
             Size of the noise map (N x N).
         noise_level: float
@@ -17,6 +18,7 @@ def generate_noise_map(image_size, noise_level, pix_size):
             size of pixels in arcminutes
 
     Returns:
+    --------
         ndarray: Noise map.
     """
     
