+++ deepszsim/noise.py
@@ -11,15 +11,16 @@ def generate_noise_map(image_size, noise_level, pix_size):
     Parameters:
     -----------
         image_size: int
-            Size of the noise map (N x N).
+            Size of the noise map with dimensions N x N
         noise_level: float
-            Noise level of the survey.
+            Noise level of the survey
         pix_size: int
             size of pixels in arcminutes
 
     Returns:
     --------
-        ndarray: Noise map.
+        noise_map: array
+            2D map of the white noise
     """
     
     # Create random noise map
@@ -28,6 +29,5 @@ def generate_noise_map(image_size, noise_level, pix_size):
     # Scale random noise map by noise level
     scaled_noise_map = random_noise_map * noise_level
     noise_map = scaled_noise_map / pix_size
-    
 
     return noise_map
