### 2025-05-25 18a1c9e
removed "raises" part of docstring; formatting
