### 2025-05-25 a4c7f86
docstrings: grammar/formatting
