+++ deepszsim/simclusters.py
@@ -10,41 +10,41 @@ from datetime import datetime as dt
 
 class simulate_clusters:
     """
-    class for simulating a distribution of clusters
+    Class for simulating a distribution of clusters.
     
     Parameters
     ----------
     M200: float or array-like of float
-        the mass contained within R200 in solar masses (same length as z_dist)
+        mass contained within R200 in solar masses (same length as z_dist).
     redshift_z: float or array-like of float
-        the redshift of the cluster (unitless) (same length as M200_dist)
+        redshift of the cluster (unitless) (same length as M200_dist).
     num_halos: None or int
-        number of halos to simulate if none supplied
+        number of halos to simulate if none supplied.
     halo_params_dict: None or dict
         parameters from which to sample halos if `num_halos` specified,
-        must contain `zmin`, `zmax`, `m200min_SM`, `m200max_SM`
+        must contain `zmin`, `zmax`, `m200min_SM`, `m200max_SM`.
     R200_Mpc: None or float or np.ndarray(float)
         if None, will calculate the R200 values corresponding to a given set of
-        M200 and redshift_z values for the specified cosmology
+        M200 and redshift_z values for the specified cosmology.
     profile: str
-        Name of profile, currently only supports "Battaglia2012"
+        Name of profile, currently only supports "Battaglia2012".
     image_size_pixels: None or int
         image size in pixels (should be odd; if even, will return images whose
-        sides are `image_size_pixels+1` in length)
+        sides are `image_size_pixels+1` in length).
     image_size_arcmin: None or float
-        image size in arcmin
+        image size in arcmin.
     pixel_size_arcmin: None or float
-        pixel size in arcmin
+        pixel size in arcmin.
     alpha: float
-        fixed to equal 1.0 in Battaglia 2012
+        fixed to equal 1.0 in Battaglia 2012.
     gamma: float
-        fixed to equal -0.3 in Battaglia 2012
+        fixed to equal -0.3 in Battaglia 2012.
     load_vars_yaml: None or str
-        path to yaml file with params; if `None`, must explicitly include image specifications
+        path to yaml file with params; if `None`, must explicitly include image specifications.
     seed: None or int
-        random seed value to sample with
+        random seed value to sample with.
     tqverb: bool
-        whether or not to display tqdm progress bar while making T maps
+        whether or not to display tqdm progress bar while making T maps.
     """
     def __init__(self, M200 = None, redshift_z = None, num_halos = None, halo_params_dict = None,
                  R200_Mpc = None, profile = "Battaglia2012",
@@ -116,7 +116,7 @@ class simulate_clusters:
     
     def get_y_maps(self):
         """
-        get 'y' maps from object
+        Get 'y' maps from object.
 
         Parameters
         ----------
@@ -124,8 +124,8 @@ class simulate_clusters:
         
         Returns
         -------
-        np.ndarray(float)
-            self._size many maps of the Compton `y` value, each of which is image_size_pixels x image_size_pixels in size
+        self.y_maps: array
+            self._size many maps of the Compton `y` value, each of which is image_size_pixels x image_size_pixels in size.
         """
         try:
             return self.y_maps
@@ -148,9 +148,9 @@ class simulate_clusters:
         
         Returns
         -------
-        np.ndarray(float)
+        self.dT_maps: array(float)
             self._size many maps of the dT values in units of uK, each of which is image_size_pixels x
-            image_size_pixels in size
+            image_size_pixels in size.
         """
         try:
             return self.dT_maps
@@ -161,21 +161,21 @@ class simulate_clusters:
             return self.dT_maps
     
     def get_T_maps(self, add_CMB = True, returnval = False):
-        """
-        get 'T' maps from object
+        """.
[…]
