### 2025-05-25 924b421
docstrings: formatting, grammar, move comments
