### 2025-05-25 a518990
docstring: formatting/grammar
