### 2025-05-25 2f55e18
docstrings: grammar
