### 2025-05-25 163f294
docstrings: gramma, formatting, func headings
