+++ deepszsim/io_params.py
@@ -76,7 +76,7 @@ class config_obj:
 
         Returns
         -------
-            none
+        none
         """
         attr_split = re.split("\.", attr)
         if attr in self._all_params_dict['USERPARAMS']:
@@ -101,7 +101,7 @@ class config_obj:
 
         Returns
         -------
-            none
+        none
         """
         with open(os.path.join(savedir, f"{run_id}_params.yaml"), permission) as yaml_file:
             dump = pyyaml.dump(self.dict, default_flow_style = False, allow_unicode = True, encoding = None)
@@ -110,7 +110,7 @@ class config_obj:
             
     def _generate_run_id(random_digits=6):
         """
-        Generates run id; typically used in filenames
+        Generates a run id; typically used in filenames
         
         Parameters
         ----------
@@ -131,7 +131,7 @@ class config_obj:
 
     def cosmology_param(self, ref):
         """
-        <>
+        Set cosmology parameters
         Parameters
         ----------
         attr : str
@@ -152,15 +152,16 @@ class config_obj:
 
     def _quick_yaml_load(infile):
         """
-        simply load yaml files without remembering the syntax or yaml.safe_load command
+        Load yaml files without remembering the syntax or yaml.safe_load command
         Parameters
         ----------
         infile : str
             path to yaml file that you wish to load
+            
         Returns
         -------
-        dict
-            a "safe load" dictionary version of infile
+        yaml.safe_load(): dict
+            "safe load" dictionary version of infile
         """
         with open(infile, "r") as f:
             return yaml.safe_load(f)
