+++ deepszsim/load_vars.py
@@ -18,7 +18,7 @@ def load_vars(file_name= "config_simACTDR5.yaml",
               cosmo_name : str = None,
               enforce_odd_pix : bool = True):
     """
-    Load variables from YAML file
+    Load variables from YAML file.
     
     Parameters
     ----------
@@ -40,7 +40,7 @@ def load_vars(file_name= "config_simACTDR5.yaml",
 
     Returns
     -------
-    dict
+    dict: dict
         a dictionary that will enable you to simulate SZ clusters with a given cosmology and set of observational
         properties.
 
@@ -113,7 +113,7 @@ def load_vars(file_name= "config_simACTDR5.yaml",
 
 def readh5(fname, fdir = None):
     """
-    Read data from h5 file
+    Read data from h5 file.
     
     Parameters
     ----------
@@ -124,7 +124,7 @@ def readh5(fname, fdir = None):
         
     Returns
     -------
-    dict
+    fmdict: dict
         dictionary that represents a single cluster (if `fname` represents a single cluster) or a nested set of
         clusters. Each cluster that is returned will itself have a dictionary with 'maps' and 'params' as keys,
         containing the output maps and the parameters that were used to make them, respectively
