+++ deepszsim/filters.py
@@ -11,7 +11,7 @@ def get_tSZ_signal_aperture_photometry(dT_map, radmax_arcmin, pixel_scale,
     
     Parameters:
     ----------
-    dT_map: array(float)
+    dT_map: float array
         map in uK
     radmax_arcmin: float
         radius of the inner aperture in arcmin
@@ -23,9 +23,9 @@ def get_tSZ_signal_aperture_photometry(dT_map, radmax_arcmin, pixel_scale,
     Returns:
     -------
     disk_mean: float
-        average value within an annulus of inner radius R
+        average value within an annulus of inner radius r
     ring_mean: float
-        average value within an annulus of outer radius sqrt(2)*R
+        average value within an annulus of outer radius sqrt(2)*r
     tSZ signal: float
         thermal SZ effect signal
     """
