+++ deepszsim/io_params.py
@@ -10,16 +10,16 @@ import os
 
 class config_obj:
     """
-    configuration object that is used to obtain power spectra
+    Configuration object that is used to obtain power spectra
 
     Attributes
     ----------
-    CAMBparams : CAMBparams instance
+    CAMBparams: CAMBparams instance
         CAMB object; returns results from CAMB
-    UserParams : dict
+    UserParams: dict
         dictionary of values that the user has specified (smaller than the corresponding
         dictionary that would be necessary to fully specify a CAMBparams instance)
-    dict_iterables : dict
+    dict_iterables: dict
         dictionary of all of the iterables that the user has specified, which will be made
         available to loop over in camb_power_spectrum.CAMBPowerSpectrum
     """
@@ -31,9 +31,9 @@ class config_obj:
 
         Parameters
         ----------
-        user_config : str
+        user_config: str
             path to yaml file that contains params the user wants to change
-        base_config : str
+        base_config: str
             path to yaml file that contains baseline cosmological parameters that reflect the best-fit
             2018 Planck cosmology and which instruct CAMB to calculate useful observables. A full list
             is available at https://camb.readthedocs.io/en/latest/model.html
@@ -69,9 +69,9 @@ class config_obj:
         
         Parameters
         ----------
-        attr : str
+        attr: str
             a key of the UserParams dictionary
-        new_val : float
+        new_val: float
             new value that you wish attr to take on
 
         Returns
@@ -90,11 +90,10 @@ class config_obj:
     def write_params_yaml_new():
         """
         write updated yaml file to disk
-        incorporate run id
             
         Parameters
         ----------
-        attr : str
+        attr: str
             a key of the UserParams dictionary
         new_val : float
             new value that you wish attr to take on
@@ -110,7 +109,7 @@ class config_obj:
             
     def _generate_run_id(random_digits=6):
         """
-        Generates a run id; typically used in filenames
+        Generates a run id
         
         Parameters
         ----------
@@ -155,7 +154,7 @@ class config_obj:
         
         Parameters
         ----------
-        infile : str
+        infile: str
             path to yaml file that you wish to load
             
         Returns
