### 2025-08-11 deepskies/DeepSZSim#242 @evavagiakis
Hi @avapolzin, thank you for these points! I'm summarizing our responses here:

**Tests:**

While the tests are automated, we agree the workflow wasn’t clear about that. We have added your suggested line to the README.

**Examples:**

We have included explicit instructions for running the notebooks as per your suggestion.
We have updated the notebooks to run without errors.

**General Repository:**
We have removed these files.
