### 2025-08-11 #7774 @avapolzin
Thank you, @evavagiakis! There are handful of remaining suggestions, which I've folded into the discussion of [#241](https://github.com/deepskies/DeepSZSim/issues/241), and I've closed [#242](https://github.com/deepskies/DeepSZSim/issues/242) as I consider those points adequately addressed.

In addition to the remaining comments in [#241](https://github.com/deepskies/DeepSZSim/issues/241), just to be able to check off the box, I wanted to ask a quick question about the author list. I see several early contributors are in the acknowledgements rather than the author list (not a problem), while some people who do not appear to have made any commits to the repository who are listed as equally contributing authors who have specifically contributed to the software. Since this is a collaboration effort, I imagine their contributions might be to pieces of code that were subsumed by `DeepSZSim` or to early iterations that predate this active repository, but I wanted to raise the point for clarification.
