### 2026-01-27 #7774 @evavagiakis
Hi @eunseongleee, @JBorrow, thank you very much for the helpful review and apologies for the delay on this as we continued to navigate members involved in this project moving on to new positions. Based on your feedback, I have implemented the following:

README updates:
- I added a direct link to the CAMB GitHub repository in the README
- I added a direct link to Colossus documentation in the README
- I added a description of the tests being performed to the README

Notebook updates:
- I defined M200 as relative to the critical density early in the demo notebook
- I edited the noise units to be labeled as uK*arcmin

Future suggestions and figure_pipeline notebook:
- We greatly appreciate this feedback, and wish to implement it in the next round of work on this project. Thank you for the ideas for refinement and expansion!
