### 2026-01-29 #7774 @avapolzin
Hi, @JBorrow (and @evavagiakis) -- I have remaining comments about basic documentation that have not been addressed. There are open issues in the DeepSZSim repo to this effect (my comment: [DeepSZSim/#241](https://github.com/deepskies/DeepSZSim/issues/241); and author broken-out issues [DeepSZSim/#253](https://github.com/deepskies/DeepSZSim/issues/253) and [DeepSZSim/#254](https://github.com/deepskies/DeepSZSim/issues/254)). These are ultimately very small changes to make, requiring just a few lines of code each, and there are many projects that have implemented both a PyPI long description and used nbsphinx (in lieu of more extensive discussion alongside the API documentation) that can be used as practical examples if necessary.
